refactor: organize files according to their primary feature
